fix(cli): Fixes toUniqueId alg

+++ elitem-cli/src/main/scala/ir/sndu/server/PeerHelper.scala
@@ -3,7 +3,7 @@ package ir.sndu.server
 import ir.sndu.server.peer.{ ApiPeer, ApiPeerType }
 
 object PeerHelper {
-  def toUniqueId(peer: ApiPeer): Long = (peer.id * Math.pow(2, 32) + peer.`type`.value).toLong
+  def toUniqueId(peer: ApiPeer): Long = peer.id.toLong * Math.pow(2, 32).toLong + peer.`type`.value
   def fromUniqueId(uniqueId: Long): ApiPeer = {
     val peerType = uniqueId & Int.MaxValue
     val peerId = uniqueId >>> 32
